# specs/002-catalog-wall-favorites/contracts/recipes.openapi.yaml
openapi: 3.1.0
info:
  title: SousChef — Recipes (catalog browse & detail)
  version: 0.2.0
  description: >
    Cook-facing recipe browse and detail. Every response is filtered by the safety wall against the
    caller's profile constraints (resolved from the X-Profile-ID header). A recipe withheld by the wall
    is indistinguishable from a non-existent one (404) so the wall never leaks existence.

components:
  parameters:
    ProfileId:
      name: X-Profile-ID
      in: header
      required: true
      schema: { type: string, minLength: 1 }
      description: Passwordless cook identity. Constraints/favorites are scoped to it; never taken from the body.
  schemas:
    Category:
      type: string
      enum: [hot_drink, cold_drink, breakfast, lunch, dinner]
    NutritionSummary:
      type: object
      required: [calories, protein_g, carbs_g, fat_g, is_approximate, servings]
      properties:
        servings: { type: integer, description: Servings these values are scaled to (the cook's default). }
        calories: { type: number }
        protein_g: { type: number }
        carbs_g: { type: number }
        fat_g: { type: number }
        is_approximate: { type: boolean, description: true when some ingredients could not be mapped. }
        unmapped_ingredient_count: { type: integer, minimum: 0, description: ingredients excluded from the totals; the UI shows coverage as (count - this) of count. }
    RecipeCard:
      type: object
      required: [id, title, category, key_ingredients]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        category: { $ref: '#/components/schemas/Category' }
        key_ingredients:
          type: array
          items: { type: string }
        image_url: { type: string, nullable: true }
    RecipeCardPage:
      type: object
      required: [items, total, page, page_size]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/RecipeCard' }
        total:
          type: integer
          minimum: 0
          description: Wall-filtered card count across the whole category (only survivors are counted).
        page: { type: integer, minimum: 1, description: 1-based index of the returned page. }
        page_size: { type: integer, minimum: 1, description: Maximum cards per page (requested/clamped). }
    RecipeDetail:
      type: object
      required: [id, title, category, servings, steps, ingredients, nutrition, allergens, is_favorite]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        category: { $ref: '#/components/schemas/Category' }
        cuisine: { type: string, nullable: true }
        total_time_minutes: { type: integer, nullable: true }
        servings: { type: integer }
        steps:
          type: array
          items: { type: string }
          description: Stored steps rendered verbatim — never rewritten or invented.
        ingredients:
          type: array
          items:
            type: object
            required: [name, raw_text]
            properties:
              name: { type: string }
              quantity: { type: number, nullable: true }
              unit: { type: string, nullable: true }
              raw_text: { type: string }
        nutrition: { $ref: '#/components/schemas/NutritionSummary' }
        allergens:
          type: array
          items: { type: string }
        is_favorite: { type: boolean }
        image_url: { type: string, nullable: true, description: The recipe's own source photo, or null (widget falls back to a placeholder). }

paths:
  /recipes:
    get:
      summary: List one page of real recipe cards in a category, filtered by the wall.
      description: >
        The wall runs over the WHOLE category first, then the compliant survivors are paged, so `total`
        is the count the cook can actually reach and no withheld recipe is ever paged into view.
      parameters:
        - $ref: '#/components/parameters/ProfileId'
        - name: category
          in: query
          required: true
          schema: { $ref: '#/components/schemas/Category' }
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
          description: 1-based page index.
        - name: page_size
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 50, default: 12 }
          description: Cards per page.
      responses:
        '200':
          description: >
            One page of compliant cards for the category. `items` MAY be empty when no recipe satisfies
            the cook's constraints (or the page is past the end) — an honest empty result, never a substitute.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecipeCardPage' }
        '400': { description: Missing/invalid X-Profile-ID or category. }
        '422': { description: page or page_size out of range. }
  /recipes/{id}:
    get:
      summary: Full stored steps + scaled nutrition for one recipe, subject to the wall.
      parameters:
        - $ref: '#/components/parameters/ProfileId'
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: The recipe detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecipeDetail' }
        '400': { description: Missing/invalid X-Profile-ID. }
        '404':
          description: >
            Not found OR withheld by the wall (a recipe violating the cook's constraints is returned as
            404 so the detail path cannot bypass the wall or reveal existence).